+++ Locales/enUS.lua
@@ -142,20 +142,9 @@ L.ShowEmpowerNameTooltip = "Show the spell name for Empowers"
 
 L.OnlyShowInterruptableCast = "Only show interruptable casts"
 
--- Name Format
-L.NameFormats = "Name Formats"
-
-L.fullName = "Full Name"
-L.lastName = "Last Name"
-L.firstName = "First Name"
-L.firstNameLastInitial = "First Name Last Initial"
-L.firstInitialLastName = "First Initial Last Name"
-
-L.fullName_Example = "Cleave Training Dummy"
-L.lastName_Example = "Dummy"
-L.firstName_Example = "Cleave"
-L.firstNameLastInitial_Example = "Cleave D."
-L.firstInitialLastName_Example = "C. Dummy"
+-- Name Text
+L.MaxCharacters = "Max Characters"
+L.MaxCharactersTooltip = "Set to 0 to show the full name."
 
 -- Widgets
 L.Widgets = "Widgets"